Add automatic offset calibration
Exposing the sensor to strong magnetic fields can leave a residual magnetization in the sensor that persists even through power cycles. The MMC5983MA has a feature to reset this magnetization, and even compute the offset of each axis.

This calibration is now part of the library, and can be performed by calling calibrate_offsets(). It is also called during begin() to ensure the sensor is in a good state.